chatbot-flow/backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
redis==5.0.0
pyjwt==2.8.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
faiss-cpu==1.7.0
sentence-transformers==2.2.0
requests==2.31.0
prometheus-fastapi-instrumentator==6.0.0
python-dotenv==1.0.0